+++ WalletClient/BlockChainInfo/Model/WalletInfo.cs
@@ -5,13 +5,13 @@ namespace WalletClient.BlockChainInfo.Model
     public class WalletInfo
     {
         [JsonProperty("balance")]
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
 
         [JsonProperty("errors")]
         public string Errors { get; set; }
 
         [JsonProperty("paytxfee")]
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
 
         [JsonProperty("connected")]
         public int Connections { get; set; }
@@ -23,7 +23,7 @@ namespace WalletClient.BlockChainInfo.Model
         public bool TestNet { get; set; }
 
         [JsonProperty("difficulty")]
-        public double Difficulty { get; set; }
+        public decimal Difficulty { get; set; }
 
         [JsonProperty("blocks")]
         public int Blocks { get; set; }
