+++ WalletClient/Shared/Model/WalletTransaction.cs
@@ -15,10 +15,10 @@ namespace WalletClient.Shared.Model
         public TransactionCategory Category { get; set; }
 
         [JsonProperty("amount")]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         [JsonProperty("fee")]
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
 
         [JsonProperty("confirmations")]
         public int? Confirmations { get; set; }
